"Ajout des fonctionnalités en SQL (COUNT, GROUP BY, HAVING) et des tests unitaires."

+++ biblio-ws/biblio-ws-model/src/main/java/oc/projet/biblio/model/entity/Ouvrage.java
@@ -11,9 +11,9 @@ public interface Ouvrage {
 
     void setTitre(String titre);
 
-    int getNbDispo();
+    Long getNbDispo();
 
-    void setNbDispo(int nbDispo);
+    void setNbDispo(Long nbDispo);
 
     Set<Exemplaire> getExemplaires();
